# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Schedurx Backend API
  version: 1.0.0
  description: >
    Clinic business API — patients, doctors, appointments, and calendar tools.
    The /tools/* paths are called directly by the Ultravox voice agent during a
    live call; /internal/* is called only by schedurx-ultravox-demo-api.
servers:
  - url: https://backend.example.test
    description: Replace with this service's public URL for deployed environments.
components:
  securitySchemes:
    ToolsBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: TOOLS_API_KEY — configured into Ultravox's tool definitions by the demo bridge.
    InternalBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: INTERNAL_API_KEY — held only by schedurx-ultravox-demo-api.
    FirebaseBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        A Firebase ID token for a Staff account. Custom claims (role, clinicId,
        doctorId) are set at onboarding time via POST /internal/staff and read
        by every /api/v1/* route — clinicId is always sourced from these claims,
        never from the request itself.
    NettuWebhookKeyAuth:
      type: apiKey
      in: header
      name: nettu-scheduler-webhook-key
      description: >
        Static key set once via `PUT /api/v1/account/webhook` on the nettu
        account (see scripts/setup-nettu-webhook.js) — nettu echoes this exact
        value back on every reminder callback. Not a bearer token.
    TwilioSignatureAuth:
      type: apiKey
      in: header
      name: X-Twilio-Signature
      description: >
        HMAC-SHA1 signature over the exact request URL + form params, verified
        against TWILIO_AUTH_TOKEN (see twilio-client.js's validateSignature).
        Twilio computes and sends this automatically; it isn't a credential a
        caller sets manually.
  schemas:
    SimpleError:
      type: object
      required: [error]
      properties:
        error:
          type: string
    StructuredToolError:
      type: object
      required: [success, error]
      properties:
        success:
          const: false
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type:
                - object
                - "null"
              additionalProperties: true
...
    CalendarSuccess:
      type: object
      required: [success, data, message]
      properties:
        success:
          const: true
        data:
          type: object
          additionalProperties: true
        message:
          type:
            - string
            - "null"
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SimpleError"
          examples:
            unauthorized:
              value:
                error: Unauthorized
...
paths:
...
  /tools/appointments/book:
    post:
      summary: Create a calendar-blocked appointment.
      security:
        - ToolsBearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ultravoxCallId:
                  type: string
                slotStart:
                  type: string
                slotEnd:
                  type:
                    - string
                    - "null"
                appointmentType:
                  type: string
                reason:
                  type:
                    - string
                    - "null"
      responses:
        "201":
          description: Appointment booked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarSuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Slot no longer available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StructuredToolError"
        "422":
          description: Missing clinic, doctor, or slotStart.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StructuredToolError"
        "500":
          description: Scheduler or internal error in structured tool format.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StructuredToolError"
  /tools/appointments/send-form:
    post:
      summary: Log the intake form URL for manual delivery.
      security:
        - ToolsBearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ultravoxCallId:
                  type: string
      responses:
        "200":
          description: Form URL logged.
          content:
            application/json:
              schema:
                type: object
                required: [delivered, note, formUrl]
                properties:
                  delivered:
                    const: false
                  note:
                    type: string
                  formUrl:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: No booked appointment in call context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /tools/calendar/slots:
    post:
      summary: Get available calendar slots for the selected doctor.
      security:
        - ToolsBearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ultravoxCallId:
                  type: string
                date:
                  type: string
                dateRange:
                  type: object
                  properties:
                    start:
                      type: string
                    end:
                      type: string
                timezone:
                  type: string
      responses:
        "200":
          description: Available slots, or a structured NO_SLOTS_AVAILABLE tool error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/CalendarSuccess"
                  - $ref: "#/components/schemas/StructuredToolError"
        "400":
          description: Invalid timezone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StructuredToolError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Missing context or scheduling validation failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StructuredToolError"
        "500":
          description: Scheduler or internal error in structured tool format.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StructuredToolError"
  /tools/calendar/reschedule:
    post:
      summary: Reschedule a calendar appointment.
      security:
        - ToolsBearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ultravoxCallId:
                  type: string
                appointmentId:
                  type: string
                newStart:
                  type: string
                newEnd:
                  type:
                    - string
                    - "null"
                reason:
                  type:
                    - string
                    - "null"
      responses:
        "200":
          description: Appointment rescheduled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarSuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: New slot no longer available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StructuredToolError"
        "422":
          description: Missing context or validation failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StructuredToolError"
        "500":
          description: Scheduler or internal error in structured tool format.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StructuredToolError"
  /tools/calendar/cancel:
    post:
      summary: Cancel a calendar appointment.
      security:
        - ToolsBearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ultravoxCallId:
                  type: string
                appointmentId:
                  type: string
                reason:
                  type:
                    - string
                    - "null"
      responses:
        "200":
          description: Appointment cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarSuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Missing context or cancellation validation failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StructuredToolError"
        "500":
          description: Scheduler or internal error in structured tool format.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StructuredToolError"

  # ── /internal/staff — bootstrap only ──────────────────────────────────────────
...